+++ weblate_schemas/test_valid.py
@@ -137,6 +137,36 @@ def test_backup_blank_url():
     )
 
 
+def test_backup_unicode_url():
+    """Test memory schema being valid."""
+    validate_schema(
+        {
+            "metadata": {
+                "version": "4.13",
+                "server": "Weblate",
+                "domain": "weblate.example.com",
+                "timestamp": "2021-11-18T18:53:54.862Z",
+            },
+            "project": {
+                "name": "Hello",
+                "slug": "hello",
+                "web": "https://example.com/Tradução%20Divine%20Divinity",
+                "instructions": "",
+                "set_language_team": False,
+                "use_shared_tm": False,
+                "contribute_shared_tm": False,
+                "access_control": 0,
+                "translation_review": False,
+                "source_review": False,
+                "enable_hooks": False,
+                "language_aliases": "",
+            },
+            "labels": [],
+        },
+        "weblate-backup.schema.json",
+    )
+
+
 def test_component():
     """Test memory schema being valid."""
     validate_schema(